refactor(backend): UserService -> UserMapService
