+++ backend/src/user.map.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { UserInfo } from '../type/channel';
+import { UserInfo } from './type/channel';
 
 @Injectable()
 export class UserMapService {
